refactor: return different built-in references based on TF version
