+++ schema/builtin_references.go
@@ -4,64 +4,21 @@
 package schema
 
 import (
-	"github.com/hashicorp/hcl-lang/lang"
+	"github.com/hashicorp/go-version"
 	"github.com/hashicorp/hcl-lang/reference"
-	"github.com/hashicorp/terraform-schema/internal/schema/refscope"
-	"github.com/zclconf/go-cty/cty"
+
+	refs_v0_12 "github.com/hashicorp/terraform-schema/internal/references/0.12"
+	refs_v1_10 "github.com/hashicorp/terraform-schema/internal/references/1.10"
 )
 
-// BuiltinReferences returns known "built-in" reference targets
+// BuiltinReferencesForVersion returns known "built-in" reference targets
 // (range-less references available within any module)
-func BuiltinReferences(modPath string) reference.Targets {
-	return reference.Targets{
-		{
-			Addr: lang.Address{
-				lang.RootStep{Name: "path"},
-				lang.AttrStep{Name: "module"},
-			},
-			ScopeId: refscope.BuiltinScope,
-			Type:    cty.String,
-			Description: lang.Markdown("The filesystem path of the module where the expression is placed\n\n" +
-				modPath),
-		},
-		{
-			Addr: lang.Address{
-				lang.RootStep{Name: "path"},
-				lang.AttrStep{Name: "root"},
-			},
-			ScopeId:     refscope.BuiltinScope,
-			Type:        cty.String,
-			Description: lang.Markdown("The filesystem path of the root module of the configuration"),
-		},
-		{
-			Addr: lang.Address{
-				lang.RootStep{Name: "path"},
-				lang.AttrStep{Name: "cwd"},
-			},
-			ScopeId: refscope.BuiltinScope,
-			Type:    cty.String,
-			Description: lang.Markdown("The filesystem path of the current working directory.\n\n" +
-				"In normal use of Terraform this is the same as `path.root`, " +
-				"but some advanced uses of Terraform run it from a directory " +
-				"other than the root module directory, causing these paths to be different."),
-		},
-		{
-			Addr: lang.Address{
-				lang.RootStep{Name: "terraform"},
-				lang.AttrStep{Name: "workspace"},
-			},
-			ScopeId:     refscope.BuiltinScope,
-			Type:        cty.String,
-			Description: lang.Markdown("The name of the currently selected workspace"),
-		},
-		{
-			Addr: lang.Address{
-				lang.RootStep{Name: "terraform"},
-				lang.AttrStep{Name: "applying"},
-			},
-			ScopeId:     refscope.BuiltinScope,
-			Type:        cty.Bool,
-			Description: lang.Markdown("True if Terraform is currently in the apply phase (including destroy mode), false otherwise"),
-		},
+func BuiltinReferencesForVersion(v *version.Version, modPath string) reference.Targets {
+	ver := v.Core()
+
+	if ver.GreaterThanOrEqual(v1_10) {
+		return refs_v1_10.BuiltinReferences(modPath)
 	}
+
+	return refs_v0_12.BuiltinReferences(modPath)
 }
